[move]: 엔티티 및 enum 위치 이동에 따른 의존성 변경
