+++ BE/backend/src/main/java/project/main/webstore/domain/order/enums/TransCondition.java
@@ -1,4 +1,4 @@
-package project.main.webstore.domain.orderHistory.enums;
+package project.main.webstore.domain.order.enums;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
