+++ BE/backend/src/main/java/project/main/webstore/domain/order/repository/OrderedItemRepository.java
@@ -3,7 +3,7 @@ package project.main.webstore.domain.order.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import project.main.webstore.domain.order.dto.OrderDBMonthlyPriceDto;
-import project.main.webstore.domain.orderHistory.controller.OrderedItem;
+import project.main.webstore.domain.order.enums.OrderedItem;
 
 import java.util.List;
