+++ BE/backend/src/main/java/project/main/webstore/domain/order/service/OrderService.java
@@ -12,9 +12,9 @@ import project.main.webstore.domain.order.dto.OrderDBMonthlyPriceDto;
 import project.main.webstore.domain.order.dto.OrderLocalDto;
 import project.main.webstore.domain.order.entity.Orders;
 import project.main.webstore.domain.order.enums.OrdersStatus;
+import project.main.webstore.domain.order.enums.TransCondition;
 import project.main.webstore.domain.order.exception.OrderExceptionCode;
 import project.main.webstore.domain.order.repository.OrderRepository;
-import project.main.webstore.domain.orderHistory.enums.TransCondition;
 import project.main.webstore.domain.users.entity.ShippingInfo;
 import project.main.webstore.domain.users.entity.User;
 import project.main.webstore.domain.users.exception.UserExceptionCode;
